Move plugin handling to Transpilation class
