+++ src/transpilation/index.ts
@@ -1,3 +1,4 @@
 export * from "./managed";
+export { Plugin } from "./plugins";
 export * from "./transpile";
 export * from "./transpiler";
